feat: finish project for submission

+++ lottery/views.py
@@ -1,23 +1,23 @@
 # IMPORTS
 import logging
+import copy
 
 from flask import Blueprint, render_template, request, flash
-from flask_login import login_required,current_user
-
-from app import db
+from flask_login import login_required, current_user
+from app import db, requires_roles
 from models import Draw, User
 
 
 # CONFIG
 lottery_blueprint = Blueprint('lottery', __name__, template_folder='templates')
 
-user = User.query.first()
-draw_key = user.draw_key
+
 
 # VIEWS
 # view lottery page
 @lottery_blueprint.route('/lottery')
 @login_required
+@requires_roles('user')
 def lottery():
 
     return render_template('lottery.html')
@@ -25,14 +25,14 @@ def lottery():
 
 @lottery_blueprint.route('/add_draw', methods=['POST'])
 @login_required
+@requires_roles('user')
 def add_draw():
     submitted_draw = ''
     for i in range(6):
         submitted_draw += request.form.get('no' + str(i + 1)) + ' '
     submitted_draw.strip()
     # create a new draw with the form data.
-    new_draw = Draw(user_id=current_user.id, draw=submitted_draw, win=False, round=0, draw_key=current_user.draw_key)  # TODO: update user_id [user_id=1 is a placeholder]
-
+    new_draw = Draw(user_id=current_user.id, draw=submitted_draw, win=False, round=0, draw_key=current_user.draw_key)
     # add the new draw to the database
     db.session.add(new_draw)
     db.session.commit()
@@ -45,16 +45,26 @@ def add_draw():
 # view all draws that have not been played
 @lottery_blueprint.route('/view_draws', methods=['POST'])
 @login_required
+@requires_roles('user')
 def view_draws():
-    # get all draws that have not been played [played=0]
-    playable_draws = Draw.query.filter_by(played=False).all()  # TODO: filter playable draws for current user
+
+    # get all draws that have not been played [played=0] and for current user only
+    playable_draws = Draw.query.filter_by(played=False, user_id=current_user.id).all()
+    # creates a list of copied draw objects which are independent of database.
+    playable_draws_copy = list(map(lambda x: copy.deepcopy(x), playable_draws))
+
+    # empty list for decrypted copied draw objects
+    decrypted_draws = []
+
     # if playable draws exist
     if len(playable_draws) != 0:
         # decrypt every draw
-        for draw in playable_draws:
-            draw.view_draw(draw_key)
+        for draw in playable_draws_copy:
+            user = User.query.filter_by(id=draw.user_id).first()
+            draw.view_draw(user.draw_key)
+            decrypted_draws.append(draw)
         # re-render lottery page with playable draws
-        return render_template('lottery.html', playable_draws=playable_draws)
+        return render_template('lottery.html', playable_draws=decrypted_draws)
     else:
         flash('No playable draws.')
         return lottery()
@@ -63,11 +73,13 @@ def view_draws():
 # view lottery results
 @lottery_blueprint.route('/check_draws', methods=['POST'])
 @login_required
+@requires_roles('user')
 def check_draws():
 
     # get played draws
-    played_draws = Draw.query.filter_by(played=True).all()  # TODO: filter played draws for current user
-
+    played_draws = Draw.query.filter_by(played=True, user_id=current_user.id).all()
+    for draw in played_draws:
+        draw.view_draw(current_user.draw_key)
     # if played draws exist
     if len(played_draws) != 0:
         return render_template('lottery.html', results=played_draws, played=True)
@@ -81,8 +93,10 @@ def check_draws():
 # delete all played draws
 @lottery_blueprint.route('/play_again', methods=['POST'])
 @login_required
+@requires_roles('user')
 def play_again():
-    delete_played = Draw.__table__.delete().where(Draw.played)  # TODO: delete played draws for current user only
+
+    delete_played = Draw.__table__.delete().where(Draw.played, Draw.user_id == current_user.id)
     db.session.execute(delete_played)
     db.session.commit()
